rename tfChain
