+++ packages/types/src/errors/index.ts
@@ -1,4 +1,4 @@
-export { ValidationError, TimeoutError, BaseError } from "./base_error";
-export { TFChainError } from "./tfChain/tfchain_error";
+export * from "./base_error";
+export { TFChainError } from "./tfchain/tfchain_error";
 export { ErrorModules } from "./modules";
-export * as TFChainErrors from "./tfChain/index";
+export * as TFChainErrors from "./tfchain/index";
