+++ packages/types/src/errors/tfchain/tft_price.ts
